# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tài liệu API E-commerce Fashion",
        "description": "Hợp đồng API giữa Vue.js Frontend và Laravel Backend",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/admin/customers": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Lấy danh sách khách hàng (có phân trang & lọc)",
                "operationId": "d4673fd730cc33cff28288a89908297d",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Từ khóa tìm kiếm (tên, email, số điện thoại)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Lọc theo trạng thái (1=active, 0=inactive)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Số bản ghi mỗi trang (mặc định: 10)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy danh sách thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Tạo khách hàng mới",
                "operationId": "3db5a45b991fcf59478b8fe071f53cb7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "phone_number",
                                    "password",
                                    "status"
                                ],
                                "properties": {
                                    "first_name": {
                                        "description": "Tên khách hàng",
                                        "type": "string",
                                        "example": "Văn A"
                                    },
                                    "last_name": {
                                        "description": "Họ khách hàng",
                                        "type": "string",
                                        "example": "Nguyễn"
                                    },
                                    "email": {
                                        "description": "Email, unique",
                                        "type": "string",
                                        "format": "email",
                                        "example": "customer@example.com"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "0901234567"
                                    },
                                    "password": {
                                        "description": "Bắt buộc khi tạo mới",
                                        "type": "string",
                                        "example": "secret123"
                                    },
                                    "status": {
                                        "description": "Trạng thái tài khoản",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Khách hàng đã được thêm thành công."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Lỗi validate dữ liệu"
                    }
                }
            }
        },
        "/api/admin/customers/search": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Tìm kiếm khách hàng theo chuỗi ký tự",
                "operationId": "9e0cb0dbc70f4225a86879ddba8442ed",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Chuỗi tìm kiếm (Tên, Email, SĐT)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/customers/{customer}": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Xem chi tiết khách hàng",
                "operationId": "80bc10168535cc87c7c16d944e668e44",
                "parameters": [
                    {
                        "name": "customer",
                        "in": "path",
                        "description": "ID của khách hàng",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy khách hàng"
                    }
                }
            },
            "put": {
                "tags": [
                    "Customers"
                ],
                "summary": "Cập nhật thông tin khách hàng",
                "operationId": "3cf72018f6eb2f9bd298351a66223213",
                "parameters": [
                    {
                        "name": "customer",
                        "in": "path",
                        "description": "ID của khách hàng cần cập nhật",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "phone_number",
                                    "status"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Văn B"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Trần"
                                    },
                                    "email": {
                                        "description": "Unique, bỏ qua customer hiện tại",
                                        "type": "string",
                                        "format": "email",
                                        "example": "new@example.com"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "0987654321"
                                    },
                                    "password": {
                                        "description": "Nullable khi update",
                                        "type": "string",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Khách hàng đã được cập nhật thành công."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy khách hàng"
                    },
                    "422": {
                        "description": "Lỗi validate dữ liệu"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Customers"
                ],
                "summary": "Xóa khách hàng",
                "operationId": "fa3b57e49208553a4e676c7738b4842c",
                "parameters": [
                    {
                        "name": "customer",
                        "in": "path",
                        "description": "ID của khách hàng cần xóa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Khách hàng đã được xóa thành công."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy khách hàng"
                    }
                }
            }
        },
...
    },
...
}